Use useAPI for gethistory and move the types to separate file
